[AIR-3025] qs: break oversized gitcmds.go by separate files
