+++ gitcmds/gitcmds.go
@@ -1,20 +1,14 @@
 package gitcmds
 
 import (
-	"bufio"
-	"encoding/json"
 	"errors"
 	"fmt"
 	"net/url"
 	"os"
-	"os/user"
 	"regexp"
-	"strconv"
 	"strings"
 
 	goGitPkg "github.com/go-git/go-git/v5"
-	"github.com/go-git/go-git/v5/config"
-	"github.com/spf13/cobra"
 	"github.com/untillpro/goutils/exec"
 	"github.com/untillpro/goutils/logger"
 	"github.com/untillpro/qs/internal/jira"
@@ -50,10 +44,9 @@ const (
 	IssuePRTtilePrefix = "Resolves issue"
 	IssueSign          = "Resolves #"
 
-	minIssueNoteLength             = 10
-	minPRTitleLength               = 8
-	minRepoNameLength              = 4
-	bashFilePerm       os.FileMode = 0644
+	minPRTitleLength              = 8
+	minRepoNameLength             = 4
+	bashFilePerm      os.FileMode = 0644
 
 	issuelineLength  = 5
 	issuelinePosOrg  = 4
@@ -115,134 +108,6 @@ func stashEntriesExist(wd string) (bool, error) {
 	return len(stashEntries) > 0, nil
 }
 
-// Upload uploads sources to git repo
-func Upload(cmd *cobra.Command, wd, currentBranch string, needToCommit bool) error {
-	if needToCommit {
-		commitMessage := cmd.Context().Value(utils.CtxKeyCommitMessage).(string)
-
-		stdout, stderr, err := new(exec.PipedExec).
-			Command(git, "add", ".").
-			WorkingDir(wd).
-			RunToStrings()
-		if err != nil {
-			logger.Verbose(stderr)
-
-			if len(stderr) > 0 {
-				return errors.New(stderr)
-			}
-
-			return fmt.Errorf("git add failed: %w", err)
-		}
-		logger.Verbose(stdout)
-
-		params := []string{"commit", "-a", mimm, commitMessage}
-
-		_, stderr, err = new(exec.PipedExec).
-			Command(git, params...).
-			WorkingDir(wd).
-			RunToStrings()
-		if strings.Contains(stderr, MsgPreCommitError) {
-			var response string
-			fmt.Println("")
-			printLn(strings.TrimSpace(stderr))
-			fmt.Print("Do you want to commit anyway(y/n)?")
-			_, _ = fmt.Scanln(&response)
-
-			if response != "y" {
-				return nil
-			}
-
-			params = append(params, "-n")
-			_, stderr, err = new(exec.PipedExec).
-				Command(git, params...).
-				WorkingDir(wd).
-				RunToStrings()
-		}
-		if err != nil {
-			logger.Verbose(stderr)
-
-			if len(stderr) > 0 {
-				return errors.New(stderr)
-			}
-
-			return fmt.Errorf("git commit failed: %w", err)
-		}
-	}
-
-	// make pull before push
-	stdout, stderr, err := new(exec.PipedExec).
-		Command(git, pull).
-		WorkingDir(wd).
-		RunToStrings()
-	if err != nil {
-		logger.Verbose(stderr)
-
-		if len(stderr) > 0 {
-			return errors.New(stderr)
-		}
-
-		return fmt.Errorf("error pulling before push: %w", err)
-	}
-
-	err = utils.Retry(func() error {
-		stdout, stderr, err = new(exec.PipedExec).
-			Command(git, push, origin, utils.RefsNotes).
-			WorkingDir(wd).
-			RunToStrings()
-		if err != nil {
-			logger.Verbose(stderr)
-
-			if len(stderr) > 0 {
-				return errors.New(stderr)
-			}
-
-			return fmt.Errorf("git push notes failed: %w", err)
-		}
-
-		return nil
-	})
-	if err != nil {
-		return err
-	}
-
-	utils.DelayIfTest()
-
-	hasUpstream, err := hasUpstreamBranch(wd, currentBranch)
-	if err != nil {
-		return fmt.Errorf("failed to check upstream branch: %w", err)
-	}
-
-	pushArgs := []string{push, origin, currentBranch}
-	if !hasUpstream {
-		pushArgs = []string{push, "-u", origin, currentBranch}
-	}
-
-	err = utils.Retry(func() error {
-		var pushErr error
-		stdout, stderr, pushErr = new(exec.PipedExec).
-			Command(git, pushArgs...).
-			WorkingDir(wd).
-			RunToStrings()
-		if pushErr != nil {
-			logger.Verbose(stderr)
-
-			if len(stderr) > 0 {
-				return errors.New(stderr)
-			}
-
-			return fmt.Errorf("git push failed: %w", pushErr)
-		}
-
-		return nil
-	})
-	if err != nil {
-		return err
-	}
-	logger.Verbose(stdout)
-
-	return nil
-}
-
 // Stash stashes uncommitted changes
 func Stash(wd string) error {
 	_, stderr, err := new(exec.PipedExec).
@@ -302,187 +167,6 @@ func HaveUncommittedChanges(wd string) (bool, error) {
 	}
 
 	return len(stdout) > 0, nil
-}
-
-// Download sources from git repo
-func Download(wd string) error {
-	// Step 1: Exit if there are uncommitted changes
-	uncommittedChanges, err := HaveUncommittedChanges(wd)
-	if err != nil {
-		return err
-	}
-
-	if uncommittedChanges {
-		return errors.New("there are uncommitted changes in the repository")
-	}
-
-	var (
-		stderr string
-		stdout string
-	)
-	// Step 2: fetch origin --prune
-	err = utils.Retry(func() error {
-		stdout, stderr, err = new(exec.PipedExec).
-			Command(git, fetch, origin, "--prune").
-			WorkingDir(wd).
-			RunToStrings()
-		if err != nil {
-			logger.Verbose(stderr)
-
-			if len(stderr) > 0 {
-				return errors.New(stderr)
-			}
-
-			return fmt.Errorf("failed to fetch origin --prune: %w", err)
-		}
-
-		return nil
-	})
-	if err != nil {
-		return err
-	}
-	logger.Verbose(stdout)
-
-	// Step 3: git fetch origin --force refs/notes/*:refs/notes/*
-	err = utils.Retry(func() error {
-		stdout, stderr, err = new(exec.PipedExec).
-			Command(git, fetch, origin, "--force", utils.RefsNotes).
-			WorkingDir(wd).
-			RunToStrings()
-		if err != nil {
-			logger.Verbose(stderr)
-
-			if len(stderr) > 0 {
-				return errors.New(stderr)
-			}
-
-			return fmt.Errorf("failed to fetch notes: %w", err)
-		}
-
-		return nil
-	})
-	if err != nil {
-		return err
-	}
-	logger.Verbose(stdout)
-
-	currentBranchName, err := GetCurrentBranchName(wd)
-	if err != nil {
-		return err
-	}
-
-	mainBranchName, err := GetMainBranch(wd)
-	if err != nil {
-		return fmt.Errorf(errMsgFailedToGetMainBranch, err)
-	}
-
-	isMain := strings.EqualFold(currentBranchName, mainBranchName)
-
-	// check out on the main branch
-	if !isMain {
-		if err := CheckoutOnBranch(wd, mainBranchName); err != nil {
-			return err
-		}
-	}
-
-	// Step 4: merge origin Main => Main with fast-forward only
-	_, stderr, err = new(exec.PipedExec).
-		Command(git, "merge", "--ff-only", fmt.Sprintf("origin/%s", mainBranchName)).
-		WorkingDir(wd).
-		RunToStrings()
-	if err != nil {
-		logger.Verbose(stderr)
-
-		// Check if fast-forward failed
-		if checkAndShowFastForwardFailure(stderr, mainBranchName) {
-			return fmt.Errorf("cannot fast-forward merge origin/%s", mainBranchName)
-		}
-
-		if len(stderr) > 0 {
-			return errors.New(stderr)
-		}
-
-		return fmt.Errorf("failed to merge origin/%s with --ff-only: %w", mainBranchName, err)
-	}
-
-	// check out back on the previous branch
-	if !isMain {
-		if err := CheckoutOnBranch(wd, currentBranchName); err != nil {
-			return err
-		}
-	}
-
-	// Step 5: If not on Main and the remote tracking branch exists merge local branch with the remote branch
-	if !isMain {
-		var hasRemoteBranch bool
-
-		hasRemoteBranch, err = hasRemoteTrackingBranch(wd, currentBranchName)
-		if err != nil {
-			return fmt.Errorf("failed to check remote tracking branch: %w", err)
-		}
-
-		if hasRemoteBranch {
-			_, stderr, err = new(exec.PipedExec).
-				Command(git, "merge", fmt.Sprintf("origin/%s", currentBranchName)).
-				WorkingDir(wd).
-				RunToStrings()
-			if err != nil {
-				logger.Verbose(stderr)
-
-				if len(stderr) > 0 {
-					return errors.New(stderr)
-				}
-
-				return fmt.Errorf("failed to merge origin/%s: %w", currentBranchName, err)
-			}
-		}
-	}
-
-	// Step 6: If upstream exists - pull upstream/Main with fast-forward only
-	upstreamExists, err := HasRemote(wd, "upstream")
-	if err != nil {
-		return fmt.Errorf("failed to check if upstream exists: %w", err)
-	}
-
-	if upstreamExists {
-		if !isMain {
-			if err := CheckoutOnBranch(wd, mainBranchName); err != nil {
-				return err
-			}
-		}
-
-		err = utils.Retry(func() error {
-			stdout, stderr, err = new(exec.PipedExec).
-				Command(git, pull, "--ff-only", "upstream", mainBranchName).
-				WorkingDir(wd).
-				RunToStrings()
-			if err != nil {
-				logger.Verbose(stderr)
-				if checkAndShowFastForwardFailure(stderr, mainBranchName) {
-					return fmt.Errorf("cannot fast-forward merge upstream/%s", mainBranchName)
-				}
-				if len(stderr) > 0 {
-					return errors.New(stderr)
-				}
-
-				return fmt.Errorf("failed to pull upstream/%s with --ff-only: %w", mainBranchName, err)
-			}
-
-			return nil
-		})
-		if err != nil {
-			return err
-		}
-		logger.Verbose(stdout)
-
-		if !isMain {
-			if err := CheckoutOnBranch(wd, currentBranchName); err != nil {
-				return err
-			}
-		}
-	}
-
-	return nil
 }
 
 func CheckoutOnBranch(wd, branchName string) error {
@@ -606,193 +290,6 @@ func GetRepoAndOrgName(wd string) (repo string, org string, err error) {
 	return
 }
 
-// ParseGitRemoteURL extracts account, repository name, and token from a git remote URL.
-// Handles HTTPS format (https://github.com/account/repo.git),
-// and HTTPS with token (https://account:token@github.com/account/repo.git or https://oauth2:token@github.com/repo.git)
-func ParseGitRemoteURL(remoteURL string) (account, repo string, token string, err error) {
-	if remoteURL == "" {
-		return "", "", "", errors.New("remote URL is empty")
-	}
-
-	remoteURL = strings.TrimSuffix(remoteURL, ".git")
-	// Handle HTTPS format: https://github.com/account/repo
-	// or https://account:token@github.com/account/repo
-	if strings.HasPrefix(remoteURL, "http") {
-		u, err := url.Parse(remoteURL)
-		if err != nil {
-			return "", "", "", fmt.Errorf("failed to parse URL: %w", err)
-		}
-
-		// Extract token if present in the userinfo section
-		if u.User != nil {
-			password, hasPassword := u.User.Password()
-			if hasPassword {
-				token = password
-			}
-		}
-
-		// Remove leading '/' if any
-		path := strings.TrimPrefix(u.Path, slash)
-		pathParts := strings.Split(path, slash)
-
-		if len(pathParts) < 2 {
-			return "", "", "", fmt.Errorf("invalid repository path in URL: %s", remoteURL)
-		}
-
-		// If no token was found or username is oauth2 (common for token auth without real account name)
-		// use the first path component as account
-		if u.User != nil && u.User.Username() == "oauth2" {
-			account = pathParts[0]
-		} else if u.User != nil && u.User.Username() != "" {
-			account = u.User.Username()
-		} else {
-			account = pathParts[0]
-		}
-
-		return account, pathParts[1], token, nil
-	}
-
-	return "", "", "", fmt.Errorf("unsupported git URL format: %s", remoteURL)
-}
-
-func IsMainOrg(wd string) (bool, error) {
-	_, org, err := GetRepoAndOrgName(wd)
-	if err != nil {
-		return false, err
-	}
-	userName, err := getUserName(wd)
-
-	return org != userName, err
-}
-
-// Fork repo
-func Fork(wd string) (string, error) {
-	repo, org, err := GetRepoAndOrgName(wd)
-	if err != nil {
-		return "", err
-	}
-
-	if len(repo) == 0 {
-		return "", errors.New(repoNotFound)
-	}
-
-	remoteURL := GetRemoteUpstreamURL(wd)
-	if len(remoteURL) > 0 {
-		return repo, errors.New(ErrAlreadyForkedMsg)
-	}
-
-	if ok, err := IsMainOrg(wd); !ok || err != nil {
-		if err != nil {
-			return repo, fmt.Errorf("IsMainOrg error: %w", err)
-		}
-
-		return repo, errors.New(ErrAlreadyForkedMsg)
-	}
-
-	_, chExist, err := ChangedFilesExist(wd)
-	if err != nil {
-		return "", err
-	}
-	if chExist {
-		stdout, stderr, err := new(exec.PipedExec).
-			Command(git, "add", ".").
-			WorkingDir(wd).
-			RunToStrings()
-		if err != nil {
-			logger.Verbose(stderr)
-
-			if len(stderr) > 0 {
-				return repo, errors.New(stderr)
-			}
-
-			return repo, fmt.Errorf("git add failed: %w", err)
-		}
-		printLn(stdout)
-
-		stdout, stderr, err = new(exec.PipedExec).
-			Command(git, "stash").
-			WorkingDir(wd).
-			RunToStrings()
-		if err != nil {
-			logger.Verbose(stderr)
-
-			if len(stderr) > 0 {
-				return repo, errors.New(stderr)
-			}
-
-			return repo, fmt.Errorf("git stash failed: %w", err)
-		}
-		printLn(stdout)
-	}
-
-	var (
-		stdout string
-		stderr string
-	)
-	err = utils.Retry(func() error {
-		stdout, stderr, err = new(exec.PipedExec).
-			Command("gh", "repo", "fork", org+slash+repo, "--clone=false").
-			WorkingDir(wd).
-			RunToStrings()
-		if err != nil {
-			logger.Verbose(stderr)
-
-			if len(stderr) > 0 {
-				return errors.New(stderr)
-			}
-
-			return fmt.Errorf("failed to fork repository: %w", err)
-		}
-
-		return nil
-	})
-	if err != nil {
-		return repo, err
-	}
-	printLn(stdout)
-
-	// Get current user name to verify fork
-	userName, err := getUserName(wd)
-	if err != nil {
-		logger.Verbose(fmt.Sprintf("Failed to get user name for verification: %v", err))
-
-		return repo, err
-	}
-
-	// Verify fork was created and is accessible with retry
-	err = utils.Retry(func() error {
-		// Try to get user email to get a valid token context, then verify repo
-		userEmail, emailErr := utils.GetUserEmail()
-		if emailErr != nil {
-			return fmt.Errorf("failed to verify GitHub authentication: %w", emailErr)
-		}
-		logger.Verbose(fmt.Sprintf("Verified GitHub authentication for user: %s", userEmail))
-
-		// Verify the forked repository exists and is accessible
-		return utils.VerifyGitHubRepoExists(userName, repo, "")
-	})
-	if err != nil {
-		logger.Verbose(fmt.Sprintf("Fork verification failed: %v", err))
-
-		return repo, fmt.Errorf("fork verification failed: %w", err)
-	}
-	_, _ = fmt.Fprintln(os.Stdout, "Fork created and verified successfully")
-
-	return repo, nil
-}
-
-func GetRemoteUpstreamURL(wd string) string {
-	stdout, _, err := new(exec.PipedExec).
-		Command(git, "config", "--local", "remote.upstream.url").
-		WorkingDir(wd).
-		RunToStrings()
-	if err != nil {
-		return ""
-	}
-
-	return strings.TrimSpace(stdout)
-}
-
 func PopStashedFiles(wd string) error {
 	if ok, err := stashEntriesExist(wd); !ok {
 		return err
@@ -848,25 +345,6 @@ func GetMainBranch(wd string) (string, error) {
 	return "", errors.New("neither main nor master branches found")
 }
 
-func getUserName(wd string) (string, error) {
-	stdout, stderr, err := new(exec.PipedExec).
-		Command("gh", "api", "user").
-		WorkingDir(wd).
-		Command("jq", "-r", ".login").
-		RunToStrings()
-	if err != nil {
-		logger.Verbose(stderr)
-
-		if len(stderr) > 0 {
-			return "", errors.New(stderr)
-		}
-
-		return "", fmt.Errorf("failed to get user name: %w", err)
-	}
-
-	return strings.TrimSpace(stdout), nil
-}
-
 func MakeUpstreamForBranch(wd string, parentRepo string) error {
 	_, stderr, err := new(exec.PipedExec).
 		Command(git, "remote", "add", "upstream", "https://github.com/"+parentRepo).
@@ -883,223 +361,6 @@ func MakeUpstreamForBranch(wd string, parentRepo string) error {
 	}
 
 	return nil
-}
-
-// MakeUpstream s.e.
-func MakeUpstream(wd string, repo string) error {
-	userName, err := getUserName(wd)
-	if err != nil {
-		return fmt.Errorf("failed to get user name: %w", err)
-	}
-
-	if len(userName) == 0 {
-		return errors.New(userNotFound)
-	}
-
-	mainBranch, err := GetMainBranch(wd)
-	if err != nil {
-		return fmt.Errorf(errMsgFailedToGetMainBranch, err)
-	}
-
-	stdout, stderr, err := new(exec.PipedExec).
-		Command(git, "remote", "rename", "origin", "upstream").
-		WorkingDir(wd).
-		RunToStrings()
-	if err != nil {
-		logger.Verbose(stderr)
-
-		if len(stderr) > 0 {
-			return errors.New(stderr)
-		}
-
-		return fmt.Errorf("failed to rename origin to upstream: %w", err)
-	}
-	printLn(stdout)
-
-	stdout, stderr, err = new(exec.PipedExec).
-		Command(git, "remote", "add", "origin", "https://github.com/"+userName+slash+repo).
-		WorkingDir(wd).
-		RunToStrings()
-	if err != nil {
-		logger.Verbose(stderr)
-
-		if len(stderr) > 0 {
-			return errors.New(stderr)
-		}
-
-		return fmt.Errorf("failed to add origin remote: %w", err)
-	}
-	printLn(stdout)
-
-	// delay to ensure remote is added
-	utils.DelayIfTest()
-
-	err = utils.Retry(func() error {
-		stdout, stderr, err = new(exec.PipedExec).
-			Command(git, "fetch", "origin").
-			WorkingDir(wd).
-			RunToStrings()
-		if err != nil {
-			logger.Verbose(stderr)
-
-			if len(stderr) > 0 {
-				return errors.New(stderr)
-			}
-
-			return fmt.Errorf("failed to fetch origin: %w", err)
-		}
-
-		return nil
-	})
-	if err != nil {
-		return err
-	}
-	logger.Verbose(stdout)
-
-	stdout, stderr, err = new(exec.PipedExec).
-		Command(git, branch, "--set-upstream-to", originSlash+mainBranch, mainBranch).
-		WorkingDir(wd).
-		RunToStrings()
-	if err != nil {
-		logger.Verbose(stderr)
-
-		if len(stderr) > 0 {
-			return errors.New(stderr)
-		}
-
-		return fmt.Errorf("failed to set upstream for main branch: %w", err)
-	}
-	printLn(stdout)
-
-	return nil
-}
-
-func GetIssueRepoFromURL(url string) (repoName string) {
-	if len(url) < 2 {
-		return
-	}
-	if strings.HasSuffix(url, slash) {
-		url = url[:len(url)-1]
-	}
-
-	arr := strings.Split(url, slash)
-	if len(arr) > issuelineLength {
-		repo := arr[len(arr)-issuelinePosRepo]
-		org := arr[len(arr)-issuelinePosOrg]
-		repoName = org + slash + repo
-	}
-
-	return
-}
-
-// CreateGithubLinkToIssue create a link between an upstream GitHub issue and the dev branch
-func CreateGithubLinkToIssue(wd, parentRepo, githubIssueURL string, issueNumber int, args ...string) (branch string, notes []string, err error) {
-	repo, org, err := GetRepoAndOrgName(wd)
-	if err != nil {
-		return "", nil, fmt.Errorf("GetRepoAndOrgName failed: %w", err)
-	}
-
-	if len(repo) == 0 {
-		return "", nil, errors.New(repoNotFound)
-	}
-
-	strIssueNum := strconv.Itoa(issueNumber)
-	myrepo := org + slash + repo
-
-	if len(args) > 0 {
-		issueURL := args[0]
-		issueRepo := GetIssueRepoFromURL(issueURL)
-		if len(issueRepo) > 0 {
-			parentRepo = issueRepo
-		}
-	}
-
-	stdout, stderr, err := new(exec.PipedExec).
-		Command("gh", "repo", "set-default", myrepo).
-		WorkingDir(wd).
-		RunToStrings()
-	if err != nil {
-		logger.Verbose(stderr)
-
-		if len(stderr) > 0 {
-			return "", nil, errors.New(stderr)
-		}
-
-		return "", nil, fmt.Errorf("failed to set default repo: %w", err)
-	}
-	printLn(stdout)
-
-	branchName, err := buildDevBranchName(githubIssueURL)
-	if err != nil {
-		return "", nil, err
-	}
-
-	// check if a branch already exists in remote
-	stdout, stderr, err = new(exec.PipedExec).
-		Command(git, "ls-remote", "--heads", origin, branch).
-		WorkingDir(wd).
-		RunToStrings()
-	if err != nil {
-		logger.Verbose(stderr)
-
-		if len(stderr) > 0 {
-			return "", nil, errors.New(stderr)
-		}
-
-		return "", nil, fmt.Errorf("failed to check if branch exists in origin remote: %w", err)
-	}
-
-	if len(stdout) > 0 {
-		return "", nil, fmt.Errorf("branch %s already exists in origin remote", branch)
-	}
-
-	mainBranch, err := GetMainBranch(wd)
-	if err != nil {
-		return "", nil, fmt.Errorf(errMsgFailedToGetMainBranch, err)
-	}
-
-	stdout, stderr, err = new(exec.PipedExec).
-		Command("gh", "issue", "develop", strIssueNum, "--branch-repo="+myrepo, "--repo="+parentRepo, "--name="+branchName, "--base="+mainBranch).
-		WorkingDir(wd).
-		RunToStrings()
-	if err != nil {
-		logger.Verbose(stderr)
-
-		if len(stderr) > 0 {
-			return "", nil, errors.New(stderr)
-		}
-
-		return "", nil, fmt.Errorf("failed to create development branch for issue: %w", err)
-	} // delay to ensure branch is created
-	logger.Verbose(stdout)
-
-	utils.DelayIfTest()
-
-	branch = strings.TrimSpace(stdout)
-	segments := strings.Split(branch, slash)
-	branch = segments[len(segments)-1]
-
-	if len(branch) == 0 {
-		return "", nil, errors.New("can not create branch for issue")
-	}
-	// old-style notes
-	issueName, err := GetIssueNameByNumber(strIssueNum, parentRepo)
-	if err != nil {
-		return "", nil, err
-	}
-
-	comment := IssuePRTtilePrefix + " '" + issueName + "' "
-	body := ""
-	if len(issueName) > 0 {
-		body = IssueSign + strIssueNum + oneSpace + issueName
-	}
-	// Prepare new notes with issue name as description
-	notesObj, err := notesPkg.Serialize(githubIssueURL, "", notesPkg.BranchTypeDev, issueName)
-	if err != nil {
-		return "", nil, err
-	}
-
-	return branch, []string{comment, body, notesObj}, nil
 }
 
 // SyncMainBranch syncs the local main branch with upstream and origin
@@ -1228,69 +489,6 @@ func GetBranchTypeByName(branchName string) notesPkg.BranchType {
 	}
 }
 
-func buildDevBranchName(issueURL string) (string, error) {
-	// Extract issue number from URL
-	parts := strings.Split(issueURL, slash)
-	if len(parts) < 2 {
-		return "", fmt.Errorf("invalid issue URL format: %s", issueURL)
-	}
-	issueNumber := parts[len(parts)-1]
-
-	// Extract owner and repo from URL
-	repoURL := strings.Split(issueURL, "/issues/")[0]
-	urlParts := strings.Split(repoURL, slash)
-	if len(urlParts) < 5 { //nolint:revive
-		return "", fmt.Errorf("invalid GitHub URL format: %s", repoURL)
-	}
-	owner := urlParts[3] //nolint:revive
-	repo := urlParts[4]  //nolint:revive
-
-	// Use gh CLI to get issue title
-	stdout, stderr, err := new(exec.PipedExec).
-		Command("gh", "issue", "view", issueNumber, "--repo", fmt.Sprintf("%s/%s", owner, repo), "--json", "title").
-		RunToStrings()
-
-	if err != nil {
-		logger.Verbose(stderr)
-
-		if len(stderr) > 0 {
-			return "", errors.New(stderr)
-		}
-
-		return "", fmt.Errorf("failed to get issue title: %w", err)
-	}
-	logger.Verbose(stdout)
-
-	// Parse JSON response
-	var issueData struct {
-		Title string `json:"title"`
-	}
-
-	if err := json.Unmarshal([]byte(stdout), &issueData); err != nil {
-		return "", fmt.Errorf("failed to parse issue data: %w", err)
-	}
-
-	// Create kebab-case version of the title
-	kebabTitle := strings.ToLower(issueData.Title)
-	// Replace spaces and special characters with dashes
-	kebabTitle = regexp.MustCompile(`[^a-z0-9]+`).ReplaceAllString(kebabTitle, "-")
-	// Remove leading and trailing dashes
-	kebabTitle = strings.Trim(kebabTitle, "-")
-
-	// Construct branch name: {issue-number}-{kebab-case-title}
-	branchName := fmt.Sprintf("%s-%s", issueNumber, kebabTitle)
-
-	// Ensure branch name doesn't exceed git's limit (usually around 250 chars)
-	if len(branchName) > maximumBranchNameLength {
-		branchName = branchName[:maximumBranchNameLength]
-	}
-	branchName = utils.CleanArgFromSpecSymbols(branchName)
-	// Add suffix "-dev" for a dev branch
-	branchName += "-dev"
-
-	return branchName, nil
-}
-
 // GetBranchType returns branch type based on notes or branch name
 func GetBranchType(wd string) (string, notesPkg.BranchType, error) {
 	currentBranchName, err := GetCurrentBranchName(wd)
@@ -1331,287 +529,6 @@ func isOldStyledBranch(notes []string) bool {
 	}
 
 	return false
-}
-
-func GetIssueNameByNumber(issueNum string, parentrepo string) (string, error) {
-	stdout, stderr, err := new(exec.PipedExec).
-		Command("gh", "issue", "view", issueNum, "--repo", parentrepo, "--json", "title").
-		RunToStrings()
-	if err != nil {
-		logger.Verbose(stderr)
-
-		if len(stderr) > 0 {
-			return "", errors.New(stderr)
-		}
-
-		return "", fmt.Errorf("failed to get issue name by number: %w", err)
-	}
-
-	type issueDetails struct {
-		Title string `json:"title"`
-	}
-
-	var issue issueDetails
-	if err := json.Unmarshal([]byte(stdout), &issue); err != nil {
-		return "", fmt.Errorf("failed to parse issue title JSON: %w", err)
-	}
-
-	return issue.Title, nil
-}
-
-// normalizeBranchName normalizes a branch name to comply with Git branch naming rules.
-// It replaces all invalid characters with dashes and ensures the name follows Git conventions.
-// Rules applied:
-// - Replace invalid characters (ASCII control chars, ~, ^, :, ?, [, *, spaces, ..) with dash
-// - Remove leading/trailing dots, slashes, and dashes
-// - Replace consecutive dashes with a single dash
-// - Ensure lowercase for consistency
-func normalizeBranchName(branchName string) string {
-	if branchName == "" {
-		return branchName
-	}
-
-	// Define characters that should be replaced with dash
-	invalidChars := []string{
-		" ", "~", "^", ":", "?", "[", "]", "*", "\\", "#", "!",
-		"\t", "\n", "\r", // whitespace characters
-		"\x00", "\x01", "\x02", "\x03", "\x04", "\x05", "\x06", "\x07", // ASCII control chars
-		"\x08", "\x09", "\x0a", "\x0b", "\x0c", "\x0d", "\x0e", "\x0f",
-		"\x10", "\x11", "\x12", "\x13", "\x14", "\x15", "\x16", "\x17",
-		"\x18", "\x19", "\x1a", "\x1b", "\x1c", "\x1d", "\x1e", "\x1f",
-		"\x7f", // DEL character
-	}
-
-	// Replace invalid characters with dash
-	normalized := branchName
-	for _, char := range invalidChars {
-		normalized = strings.ReplaceAll(normalized, char, "-")
-	}
-
-	// Replace ".." (double dot) with single dash
-	normalized = strings.ReplaceAll(normalized, "..", "-")
-
-	// Remove leading dots
-	normalized = strings.TrimLeft(normalized, ".")
-
-	// Remove trailing slashes
-	normalized = strings.TrimRight(normalized, "/")
-
-	// Remove trailing dash
-	normalized = strings.TrimRight(normalized, "-")
-
-	// Remove leading and trailing dashes and underscores
-	normalized = strings.Trim(normalized, "-_")
-
-	// Replace consecutive dashes with a single dash
-	re := regexp.MustCompile(`-+`)
-	normalized = re.ReplaceAllString(normalized, "-")
-
-	return normalized
-}
-
-// CreateDevBranch creates dev branch and pushes it to origin
-// Parameters:
-// branch - branch name
-// notes - notes for branch
-// checkRemoteBranchExistence - if true, checks if a branch already exists in remote
-func CreateDevBranch(wd, branchName, mainBranch string, notes []string, checkRemoteBranchExistence bool) error {
-	branchName = normalizeBranchName(branchName)
-	if branchName == "" {
-		return errors.New("branch name is empty after normalization")
-	}
-
-	stdout, stderr, err := new(exec.PipedExec).
-		Command(git, "checkout", mainBranch).
-		WorkingDir(wd).
-		RunToStrings()
-
-	if err != nil {
-		if strings.Contains(err.Error(), err128) && strings.Contains(stderr, "matched multiple") {
-			err = new(exec.PipedExec).
-				Command(git, "checkout", "--track", originSlash+mainBranch).
-				WorkingDir(wd).
-				Run(os.Stdout, os.Stdout)
-			if err != nil {
-				return err
-			}
-		}
-	}
-	if err != nil {
-		return err
-	}
-
-	if checkRemoteBranchExistence {
-		// check if a branch already exists in remote
-		stdout, stderr, err := new(exec.PipedExec).
-			Command(git, "ls-remote", "--heads", "origin", branchName).
-			WorkingDir(wd).
-			RunToStrings()
-		if err != nil {
-			logger.Verbose(stderr)
-
-			return err
-		}
-		logger.Verbose(stdout)
-
-		if len(stdout) > 0 {
-			return fmt.Errorf("branch %s already exists in origin remote", branchName)
-		}
-	}
-
-	// Create new branch from main
-	err = new(exec.PipedExec).
-		Command(git, "checkout", "-B", branchName).
-		WorkingDir(wd).
-		Run(os.Stdout, os.Stdout)
-	if err != nil {
-		return err
-	}
-
-	// Fetch notes from origin before pushing
-	stdout, stderr, err = new(exec.PipedExec).
-		Command(git, fetch, origin, "--force", utils.RefsNotes).
-		WorkingDir(wd).
-		RunToStrings()
-	if err != nil {
-		logger.Verbose(stderr)
-
-		return fmt.Errorf("failed to fetch notes: %w, stdout: %s", err, stdout)
-	}
-
-	// Add empty commit to for keeping notes
-	err = new(exec.PipedExec).
-		Command(git, "commit", "--allow-empty", "-m", MsgCommitForNotes).
-		WorkingDir(wd).
-		Run(os.Stdout, os.Stdout)
-	if err != nil {
-		return err
-	}
-	// Link notes to it
-	if err := AddNotes(wd, notes); err != nil {
-		return err
-	}
-
-	// Push notes to origin with retry
-	err = utils.Retry(func() error {
-		stdout, stderr, err = new(exec.PipedExec).
-			Command(git, push, origin, utils.RefsNotes).
-			WorkingDir(wd).
-			RunToStrings()
-
-		return err
-	})
-	if err != nil {
-		logger.Verbose(stderr)
-
-		return fmt.Errorf("failed to push notes to origin: %w", err)
-	}
-	utils.DelayIfTest()
-
-	// Push branch to origin with retry
-	err = utils.Retry(func() error {
-		stdout, stderr, err = new(exec.PipedExec).
-			Command(git, push, "-u", origin, branchName).
-			WorkingDir(wd).
-			RunToStrings()
-
-		return err
-	})
-	if err != nil {
-		logger.Verbose(stderr)
-
-		return fmt.Errorf("failed to push branch to origin: %w, stdout: %s", err, stdout)
-	}
-
-	utils.DelayIfTest()
-
-	return nil
-}
-
-func AddNotes(wd string, notes []string) error {
-	if len(notes) == 0 {
-		return nil
-	}
-	// Add new Notes
-	for _, s := range notes {
-		str := strings.TrimSpace(s)
-		if len(str) > 0 {
-			stdout, stderr, err := new(exec.PipedExec).
-				Command(git, "notes", "append", "-m", str).
-				WorkingDir(wd).
-				RunToStrings()
-			if err != nil {
-				logger.Verbose(stderr)
-
-				if len(stderr) > 0 {
-					return errors.New(stderr)
-				}
-
-				return fmt.Errorf("failed to add note: %w", err)
-			}
-			printLn(stdout)
-		}
-	}
-
-	return nil
-}
-
-// GetNotes returns notes for a branch
-// Returns:
-// - notes
-// - revision count
-// - error if any
-func GetNotes(wd, branchName string) (notes []string, revCount int, err error) {
-	mainBranchName, err := GetMainBranch(wd)
-	if err != nil {
-		return nil, 0, fmt.Errorf(errMsgFailedToGetMainBranch, err)
-	}
-
-	return getNotesWithMainBranch(wd, branchName, mainBranchName)
-}
-
-func getNotesWithMainBranch(wd, branchName, mainBranchName string) (notes []string, revCount int, err error) {
-	stdout, stderr, err := new(exec.PipedExec).
-		Command(git, "rev-list", mainBranchName+".."+branchName).
-		WorkingDir(wd).
-		RunToStrings()
-	if err != nil {
-		logger.Verbose(stderr)
-
-		return notes, 0, fmt.Errorf("failed to get commit list: %w", err)
-	}
-	if len(stdout) == 0 {
-		return notes, 0, errors.New("error: No commits found in current branch")
-	}
-
-	revList := strings.Split(strings.TrimSpace(stdout), caret)
-	for _, rev := range revList {
-		stdout, stderr, err := new(exec.PipedExec).
-			Command(git, "notes", "show", rev).
-			WorkingDir(wd).
-			RunToStrings()
-		if err != nil {
-			if strings.Contains(stderr, "no note found") {
-				continue
-			}
-			logger.Verbose(stderr)
-
-			return notes, len(revList), fmt.Errorf("failed to get notes: %w", err)
-		}
-		rawNotes := strings.Split(stdout, caret)
-		for _, rawNote := range rawNotes {
-			note := strings.TrimSpace(rawNote)
-			if len(note) > 0 {
-				notes = append(notes, note)
-			}
-		}
-	}
-
-	if len(notes) == 0 {
-		return notes, len(revList), errors.New("error: No notes found in current branch")
-	}
-
-	return notes, len(revList), nil
 }
 
 // GetParentRepoName - parent repo of forked
@@ -1784,96 +701,6 @@ func PullUpstream(wd string) error {
 	return nil
 }
 
-// GetGoneBranchesLocal returns gone local branches
-func GetGoneBranchesLocal(wd string) (*[]string, error) {
-	// https://dev.heeus.io/launchpad/#!14544
-	// 1. Step
-	_, _, err := new(exec.PipedExec).
-		Command(git, fetch, "-p", "--dry-run").
-		WorkingDir(wd).
-		RunToStrings()
-	if err != nil {
-		return nil, err
-	}
-	_, _, err = new(exec.PipedExec).
-		Command(git, fetch, "-p").
-		WorkingDir(wd).
-		RunToStrings()
-	if err != nil {
-		return nil, err
-	}
-	// 2. Step
-	stdout, _, err := new(exec.PipedExec).
-		Command(git, branch, "-vv").
-		WorkingDir(wd).
-		Command("grep", "\\[[^]]*: gone[^]]*\\]").
-		Command("gawk", "{print $1}").
-		RunToStrings()
-	if err != nil {
-		return nil, err
-	}
-	// alternate: grep '\[.*: gone'
-	mbrlocallist := strings.Split(stdout, caret)
-
-	stsr := []string{}
-	stdout, _, err = new(exec.PipedExec).
-		Command(git, branch, "-r").
-		WorkingDir(wd).
-		RunToStrings()
-	if err != nil {
-		return nil, err
-	}
-	myremotelist := strings.Split(stdout, caret)
-	mainBranch, err := GetMainBranch(wd)
-	if err != nil {
-		return nil, fmt.Errorf(errMsgFailedToGetMainBranch, err)
-	}
-
-	curbr, err := GetCurrentBranchName(wd)
-	if err != nil {
-		return nil, err
-	}
-
-	for _, mylocalbranch := range mbrlocallist {
-		mybranch := strings.TrimSpace(mylocalbranch)
-		bfound := false
-		if strings.Contains(mybranch, curbr) {
-			bfound = true
-		} else if !strings.Contains(mybranch, mainBranch) && !strings.Contains(mybranch, "HEAD") {
-			for _, mbranch := range myremotelist {
-				mbranch = strings.TrimSpace(mbranch)
-				if mybranch == mbranch {
-					bfound = true
-					break
-				}
-			}
-		}
-		if !bfound {
-			// delete branch in fork
-			stsr = append(stsr, mybranch)
-		}
-	}
-	return &stsr, nil
-}
-
-// DeleteBranchesLocal s.e.
-func DeleteBranchesLocal(wd string, strs *[]string) error {
-	for _, str := range *strs {
-		if strings.TrimSpace(str) != "" {
-			_, _, err := new(exec.PipedExec).
-				Command(git, branch, "-D", str).
-				WorkingDir(wd).
-				RunToStrings()
-			fmt.Printf("Branch %s deleted\n", str)
-			if err != nil {
-				return err
-			}
-		}
-	}
-
-	return nil
-}
-
 func GetNoteAndURL(notes []string) (note string, url string) {
 	for _, s := range notes {
 		s = strings.TrimSpace(s)
@@ -1898,154 +725,6 @@ func GetNoteAndURL(notes []string) (note string, url string) {
 	return note, url
 }
 
-func GetBodyFromNotes(notes []string) string {
-	b := ""
-	if (len(notes) > 1) && strings.Contains(strings.ToLower(notes[0]), strings.ToLower(IssuePRTtilePrefix)) {
-		for i, note := range notes {
-			note = strings.TrimSpace(note)
-			if (strings.Contains(note, "https://") && !strings.Contains(note, "/issues/")) || !strings.Contains(note, "https://") {
-				strings.Split(strings.ReplaceAll(note, "\r\n", caret), "")
-				if i > 0 && len(note) > 0 {
-					b += note
-				}
-			}
-		}
-	}
-	return b
-}
-
-func createPR(
-	wd,
-	parentRepoName,
-	prBranchName,
-	issueDescription string,
-	notes []string,
-	asDraft bool,
-) (stdout string, stderr string, err error) {
-	if len(notes) == 0 {
-		return "", "", errors.New(ErrMsgPRNotesImpossible)
-	}
-
-	var isCustomBranch bool
-	prTitle := issueDescription
-	if len(prTitle) == 0 {
-		isCustomBranch = true
-		fmt.Print("Enter pull request title: ")
-		reader := bufio.NewReader(os.Stdin)
-
-		// Read until newline (includes spaces)
-		prTitle, err = reader.ReadString(caretByte)
-		if err != nil {
-			return "", "", err
-		}
-
-		prTitle = strings.TrimSpace(prTitle)
-		if len(prTitle) < minPRTitleLength {
-			return "", "", errors.New("too short pull request title")
-		}
-	}
-
-	var strNotes string
-	var issueURL string
-	strNotes, issueURL = GetNoteAndURL(notes)
-	b := GetBodyFromNotes(notes)
-	if len(b) == 0 && !isCustomBranch {
-		b = strNotes
-	}
-	if len(issueURL) > 0 {
-		b = b + caret + issueURL
-	}
-	strBody := fmt.Sprintln(b)
-
-	repoName, forkAccount, err := GetRepoAndOrgName(wd)
-	if err != nil {
-		return "", "", err
-	}
-
-	repo := parentRepoName
-	if len(repo) == 0 {
-		// Single remote mode: no parent repo, PR to same repo
-		repo = forkAccount + slash + repoName
-	}
-
-	// Determine the head reference for the PR
-	// In fork mode: use "forkAccount:branchName" format
-	// In single remote mode: use just "branchName" format
-	headRef := prBranchName
-	if len(parentRepoName) > 0 {
-		// Fork mode: specify the fork account
-		headRef = forkAccount + ":" + prBranchName
-	}
-
-	args := []string{
-		"pr",
-		"create",
-		fmt.Sprintf(`--head=%s`, headRef),
-		fmt.Sprintf(`--repo=%s`, repo),
-		fmt.Sprintf(`--body=%s`, strings.TrimSpace(strBody)),
-		fmt.Sprintf(`--title=%s`, strings.TrimSpace(prTitle)),
-	}
-	if asDraft {
-		args = append(args, "--draft")
-	}
-	err = utils.Retry(func() error {
-		stdout, stderr, err = new(exec.PipedExec).
-			Command("gh", args...).
-			RunToStrings()
-
-		return err
-	})
-	if err != nil {
-		return stdout, stderr, err
-	}
-
-	prInfo, stdout, stderr, err := DoesPrExist(wd, parentRepoName, prBranchName, PRStateOpen)
-	if err != nil {
-		return stdout, stderr, err
-	}
-	if prInfo == nil {
-		return stdout, stderr, errors.New("PR not created")
-	}
-	// print PR URL
-	if len(prInfo.URL) > 0 {
-		fmt.Println(prInfo.URL)
-	}
-
-	return stdout, stderr, err
-}
-
-func retrieveRepoNameFromUPL(prurl string) string {
-	var strs = strings.Split(prurl, slash)
-	if len(strs) < minRepoNameLength {
-		return ""
-	}
-	res := ""
-	lenstr := len(strs)
-	for i := lenstr - minRepoNameLength; i < lenstr-2; i++ {
-		if res == "" {
-			res = strs[i]
-		} else {
-			res = res + slash + strs[i]
-		}
-	}
-	return res
-}
-
-// getRemotes shows list of names of all remotes
-func getRemotes(wd string) []string {
-	stdout, _, _ := new(exec.PipedExec).
-		Command(git, "remote").
-		WorkingDir(wd).
-		RunToStrings()
-	strs := strings.Split(stdout, caret)
-	for i, str := range strs {
-		if len(strings.TrimSpace(str)) == 0 {
-			strs = append(strs[:i], strs[i+1:]...)
-		}
-	}
-	return strs
-}
-
 // GetFilesForCommit shows list of file names, ready for commit
 func GetFilesForCommit(wd string) []string {
 	stdout, _, _ := new(exec.PipedExec).
@@ -2060,172 +739,6 @@ func GetFilesForCommit(wd string) []string {
 		}
 	}
 	return strs
-}
-
-// hasUpstreamBranch checks if the current branch has an upstream tracking branch configured
-func hasUpstreamBranch(wd string, branchName string) (bool, error) {
-	stdout, _, err := new(exec.PipedExec).
-		Command(git, "config", "--get", fmt.Sprintf("branch.%s.remote", branchName)).
-		WorkingDir(wd).
-		RunToStrings()
-
-	if err != nil {
-		// If the config doesn't exist, git config returns exit code 1
-		// This is expected when no upstream is configured
-		return false, nil
-	}
-
-	return strings.TrimSpace(stdout) != "", nil
-}
-
-// hasRemoteTrackingBranch checks if a remote tracking branch exists for the given branch
-func hasRemoteTrackingBranch(wd string, branchName string) (bool, error) {
-	stdout, stderr, err := new(exec.PipedExec).
-		Command(git, "branch", "-r").
-		WorkingDir(wd).
-		Command("grep", branchName).
-		RunToStrings()
-	if len(stdout) == 0 {
-		return false, nil
-	}
-
-	if err != nil {
-		logger.Verbose(stderr)
-
-		return false, err
-	}
-
-	strBranches := strings.TrimSpace(stdout)
-	if strBranches == "" {
-		return false, nil
-	}
-
-	rBranches := strings.Split(strBranches, "\n")
-	for _, rBranch := range rBranches {
-		if strings.TrimSpace(rBranch) == "origin/"+branchName {
-			return true, nil
-		}
-	}
-
-	return false, nil
-}
-
-func getGlobalHookFolder() string {
-	stdout, _, err := new(exec.PipedExec).
-		Command(git, "config", "--global", "core.hooksPath").
-		RunToStrings()
-	if err != nil {
-		return ""
-	}
-	return strings.TrimSpace(stdout)
-}
-
-func getLocalHookFolder(wd string) (string, error) {
-	dir, err := GetRootFolder(wd)
-	if err != nil {
-		return "", err
-	}
-	filename := "/.git/hooks/pre-commit"
-	filepath := dir + filename
-
-	return strings.TrimSpace(filepath), nil
-}
-
-// GlobalPreCommitHookExist - s.e.
-func GlobalPreCommitHookExist() (bool, error) {
-	filepath := getGlobalHookFolder()
-	if len(filepath) == 0 {
-		return false, nil // global hook folder not defined
-	}
-	err := os.MkdirAll(filepath, os.ModePerm)
-	if err != nil {
-		return false, err
-	}
-
-	filepath += "/pre-commit"
-	// Check if the file already exists
-	if _, err := os.Stat(filepath); os.IsNotExist(err) {
-		return false, nil // File pre-commit does not exist
-	}
-
-	return largeFileHookExist(filepath), nil
-}
-
-// LocalPreCommitHookExist - s.e.
-func LocalPreCommitHookExist(wd string) (bool, error) {
-	filepath, err := getLocalHookFolder(wd)
-	if err != nil {
-		return false, err
-	}
-	// Check if the file already exists
-	if _, err := os.Stat(filepath); os.IsNotExist(err) {
-		return false, nil
-	}
-
-	return largeFileHookExist(filepath), nil
-}
-
-func largeFileHookExist(filepath string) bool {
-	substring := LargeFileHookFilename
-	_, _, err := new(exec.PipedExec).Command("grep", "-l", substring, filepath).RunToStrings()
-
-	return err == nil
-}
-
-// SetGlobalPreCommitHook - s.e.
-func SetGlobalPreCommitHook(wd string) error {
-	var err error
-	path := getGlobalHookFolder()
-
-	if len(path) == 0 {
-		rootUser, err := user.Current()
-		if err != nil {
-			return err
-		}
-
-		path = rootUser.HomeDir
-		path += "/.git/hooks"
-		if err = os.MkdirAll(path, os.ModePerm); err != nil {
-			return err
-		}
-	}
-
-	// Set global hooks folder
-	err = new(exec.PipedExec).
-		Command(git, "config", "--global", "core.hookspath", path).
-		Run(os.Stdout, os.Stdout)
-	if err != nil {
-		return err
-	}
-
-	filepath := path + "/pre-commit"
-	f, err := createOrOpenFile(filepath)
-	if err != nil {
-		return err
-	}
-
-	_ = f.Close()
-	if !largeFileHookExist(filepath) {
-		return fillPreCommitFile(wd, filepath)
-	}
-
-	return nil
-}
-
-func GetRootFolder(wd string) (string, error) {
-	stdout, _, err := new(exec.PipedExec).
-		Command(git, "rev-parse", "--show-toplevel").
-		WorkingDir(wd).
-		RunToStrings()
-	if err != nil {
-		return "", err
-	}
-
-	return strings.TrimSpace(stdout), nil
-}
-
-func UpstreamNotExist(wd string) bool {
-	return len(getRemotes(wd)) < 2
 }
 
 func HasRemote(wd, remoteName string) (bool, error) {
@@ -2253,94 +766,6 @@ func HasRemote(wd, remoteName string) (bool, error) {
 	return false, nil
 }
 
-// GetEffectiveUpstreamRemote returns the remote to use as upstream.
-// If 'upstream' remote exists, returns "upstream".
-// Otherwise, returns "origin" (single remote mode).
-// This allows qs to work in both fork and non-fork workflows.
-func GetEffectiveUpstreamRemote(wd string) (string, error) {
-	upstreamExists, err := HasRemote(wd, "upstream")
-	if err != nil {
-		return "", err
-	}
-
-	if upstreamExists {
-		return "upstream", nil
-	}
-
-	return "origin", nil
-}
-
-func extractIntegerPrefix(input string) (string, error) {
-	// Define the regular expression pattern
-	pattern := `^\d+`
-	re := regexp.MustCompile(pattern)
-
-	// Find the match
-	match := re.FindString(input)
-	if match == "" {
-		return "", fmt.Errorf("no integer found at the beginning of the string")
-	}
-
-	// Convert the matched string to an integer
-	integerValue, err := strconv.Atoi(match)
-	if err != nil {
-		return "", fmt.Errorf("error converting string to integer: %w", err)
-	}
-
-	return strconv.Itoa(integerValue), nil
-}
-
-func issuenumExists(parentrepo string, issuenum string) bool {
-	stdouts, _, err := new(exec.PipedExec).
-		Command("gh", "issue", "develop", issuenum, "--list", "-R", parentrepo).
-		Command("gawk", "{print $2}").
-		RunToStrings()
-	if (err == nil) && (len(stdouts) > minIssueNoteLength) {
-		names := strings.Split(stdouts, caret)
-		for _, name := range names {
-			if strings.Contains(name, slash+issuenum+"-") {
-				return true
-			}
-		}
-	}
-	return false
-}
-
-func GetIssueNumFromBranchName(parentrepo string, curbranch string) (issuenum string, ok bool) {
-
-	tempissuenum, err := extractIntegerPrefix(curbranch)
-	if tempissuenum == "" {
-		return "", false
-	}
-	if err == nil {
-		if issuenumExists(parentrepo, tempissuenum) {
-			return tempissuenum, true
-		}
-	}
-
-	stdouts, stderr, err := new(exec.PipedExec).
-		Command("gh", "issue", "list", "-R", parentrepo).
-		Command("gawk", "{print $1}").
-		RunToStrings()
-	if err != nil {
-		logger.Verbose("GetIssueNumFromBranchName:", stderr)
-		return "", false
-	}
-	issuenums := strings.Split(stdouts, caret)
-	fmt.Println("Searching linked issue ")
-
-	for _, issuenum := range issuenums {
-		if len(issuenum) > 0 {
-			fmt.Println("  Issue number: ", issuenum, "...")
-			if issuenumExists(parentrepo, issuenum) {
-				return issuenum, true
-			}
-		}
-	}
-
-	return "", false
-}
-
 func GetCurrentBranchName(wd string) (string, error) {
 	branchName, stderr, err := new(exec.PipedExec).
 		Command(git, branch, "--show-current").
@@ -2357,36 +782,6 @@ func GetCurrentBranchName(wd string) (string, error) {
 	}
 
 	return strings.TrimSpace(branchName), nil
-}
-
-// createRemote creates a remote in the cloned repository
-func CreateRemote(wd, remote, account, token, repoName string, isUpstream bool) error {
-	repo, err := OpenGitRepository(wd)
-	if err != nil {
-		return err
-	}
-
-	if err = repo.DeleteRemote(remote); err != nil {
-		if !errors.Is(err, goGitPkg.ErrRemoteNotFound) {
-			return fmt.Errorf("failed to delete %s remote: %w", remote, err)
-		}
-	}
-
-	remoteURL := BuildRemoteURL(account, token, repoName, isUpstream)
-	_, err = repo.CreateRemote(&config.RemoteConfig{
-		Name: remote,
-		URLs: []string{remoteURL},
-	})
-	if err != nil {
-		return fmt.Errorf("failed to create %s remote: %w", remote, err)
-	}
-
-	return nil
-}
-
-// buildRemoteURL constructs the remote URL for cloning
-func BuildRemoteURL(account, token, repoName string, isUpstream bool) string {
-	return "https://" + account + ":" + token + "@github.com/" + account + slash + repoName + ".git"
 }
 
 // IamInMainBranch checks if current branch is main branch
@@ -2532,4 +927,53 @@ func RemoveBranch(wd, branchName string) error {
 
 		return nil
 	})
+}
+
+// ParseGitRemoteURL extracts account, repository name, and token from a git remote URL.
+// Handles HTTPS format (https://github.com/account/repo.git),
+// and HTTPS with token (https://account:token@github.com/account/repo.git or https://oauth2:token@github.com/repo.git)
+func ParseGitRemoteURL(remoteURL string) (account, repo string, token string, err error) {
+	if remoteURL == "" {
+		return "", "", "", errors.New("remote URL is empty")
+	}
+
+	remoteURL = strings.TrimSuffix(remoteURL, ".git")
+	// Handle HTTPS format: https://github.com/account/repo
+	// or https://account:token@github.com/account/repo
+	if strings.HasPrefix(remoteURL, "http") {
+		u, err := url.Parse(remoteURL)
+		if err != nil {
+			return "", "", "", fmt.Errorf("failed to parse URL: %w", err)
+		}
+
+		// Extract token if present in the userinfo section
+		if u.User != nil {
+			password, hasPassword := u.User.Password()
+			if hasPassword {
+				token = password
+			}
+		}
+
+		// Remove leading '/' if any
+		path := strings.TrimPrefix(u.Path, slash)
+		pathParts := strings.Split(path, slash)
+
+		if len(pathParts) < 2 {
+			return "", "", "", fmt.Errorf("invalid repository path in URL: %s", remoteURL)
+		}
+
+		// If no token was found or username is oauth2 (common for token auth without real account name)
+		// use the first path component as account
+		if u.User != nil && u.User.Username() == "oauth2" {
+			account = pathParts[0]
+		} else if u.User != nil && u.User.Username() != "" {
+			account = u.User.Username()
+		} else {
+			account = pathParts[0]
+		}
+
+		return account, pathParts[1], token, nil
+	}
+
+	return "", "", "", fmt.Errorf("unsupported git URL format: %s", remoteURL)
 }
